[Issue 3246] fix: update balance calculation for assets pallet
